+++ tests/unit/exportplan/test_helpers.py
@@ -327,29 +327,54 @@ def test_target_market_documents_delete(mock_target_market_documents_delete):
 
 
 @pytest.mark.django_db
-def test_get_all_lesson_details(curated_list_pages_with_lessons):
-    lessons = helpers.get_all_lesson_details()
+def test_get_lesson_details(curated_list_pages_with_lessons):
+    lesson_list = ['lesson-a1', 'lesson-a2', 'lesson-b1']
+    lessons = helpers.get_lesson_details(lesson_list)
     assert lessons == {
         'lesson-a1': {
-            'topic_name': 'Some title',
+            'category': 'Some title',
             'title': 'Lesson A1',
-            'estimated_read_duration': '2 hour 45 min',
+            'read_duration': '2 hour 45 min',
             'url': None,
         },
         'lesson-a2': {
-            'topic_name': 'Some title',
+            'category': 'Some title',
             'title': 'Lesson A2',
-            'estimated_read_duration': '12 min',
+            'duration': '12 min',
             'url': None,
         },
         'lesson-b1': {
-            'topic_name': 'Some title b',
+            'category': 'Some title b',
             'title': 'Lesson b1',
-            'estimated_read_duration': '10 min',
+            'duration': '10 min',
             'url': None,
         },
     }
 
+@pytest.mark.django_db
+def test_get_lesson_details_empty(curated_list_pages_with_lessons):
+    lesson_list = []
+    lessons = helpers.get_lesson_details(lesson_list)
+    assert lessons == {
+        'lesson-a1': {
+            'category': 'Some title',
+            'title': 'Lesson A1',
+            'read_duration': '2 hour 45 min',
+            'url': None,
+        },
+        'lesson-a2': {
+            'category': 'Some title',
+            'title': 'Lesson A2',
+            'duration': '12 min',
+            'url': None,
+        },
+        'lesson-b1': {
+            'category': 'Some title b',
+            'title': 'Lesson b1',
+            'duration': '10 min',
+            'url': None,
+        },
+    }
 
 @pytest.mark.parametrize(
     'export_plan_data, expected',